refactor: merged profile and dashboard page
