+++ lms/lms/doctype/lms_course/lms_course.py
@@ -12,10 +12,12 @@ from lms.lms.utils import get_chapters
 
 class LMSCourse(Document):
 
+
     def validate(self):
         self.validate_instructors()
         self.validate_status()
         self.image = validate_image(self.image)
+
 
     def validate_instructors(self):
         if self.is_new() and not self.instructors:
