+++ lms/patches/v0_0/set_dashboard.py
@@ -1,0 +1,4 @@
+import frappe
+
+def execute():
+    frappe.db.set_value("Portal Settings", None, "default_portal_home", "/users")
