+++ lms/templates/certificates_section.html
@@ -5,7 +5,7 @@
     {% for certificate in certificates %}
         {% set course = frappe.db.get_value("LMS Course", certificate.course, ["title", "name", "image"], as_dict=True) %}
 
-        <div class="common-card-style column-card">
+        <div class="common-card-style column-card medium">
             <div class="font-weight-bold">
                 {{ course.title }}
             </div>
