+++ lms/templates/search_course/search_course.html
@@ -1,13 +1,7 @@
 {% set search_placeholder = frappe.db.get_single_value("LMS Settings", "search_placeholder") %}
 {% set portal_course_creation = frappe.db.get_single_value("LMS Settings", "portal_course_creation") %}
 
-<div class="course-search-header">
-    <input class="search" id="search-course" placeholder="{{ _(search_placeholder) or 'Search' }}">
-    {% if portal_course_creation == "Anyone" or has_course_instructor_role() %}
-    <a class="btn btn-secondary btn-md ml-2" href="/courses/new-course"> {{ _("Create a Course") }} </a>
-    {% endif %}
-</div>
-
+<input class="search" id="search-course" placeholder="{{ _(search_placeholder) or 'Search' }}">
 <div class="empty-state alert alert-dismissible search-empty-state hide">
     <a href="#" class="close-search-empty-state" aria-label="close">&times;</a>
     <div>
