--- lms/www/dashboard/index.html
@@ -1,75 +1,0 @@
-{% extends "templates/base.html" %}
-{% block title %}
-{{ _("Dashboard")}}
-{% endblock %}
-
-
-{% block content %}
-
-<div class="common-page-style dashboard">
-    <div class="container">
-
-        {% if show_creators_section %}
-        <a class="btn btn-secondary btn-sm course-creation-link" id="create-course-link" href="/courses/new-course">
-            {{ _("Create a Course") }}
-        </a>
-        {% endif %}
-
-        <ul class="nav lms-nav" id="courses-tab">
-            <li class="nav-item">
-                <a class="nav-link active" data-toggle="tab" href="#courses-enrolled">
-                    {{ _("Enrolled") }}
-                </a>
-            </li>
-
-            {% if show_creators_section %}
-            <li class="nav-item">
-                <a class="nav-link" data-toggle="tab" href="#courses-created">
-                    {{ _("Created") }}
-                </a>
-            </li>
-            {% endif %}
-
-            {% if show_review_section %}
-            <li class="nav-item">
-                <a class="nav-link" data-toggle="tab" href="#courses-under-review">
-                    {{ _("Under Review") }}
-                </a>
-            </li>
-            {% endif %}
-            <li class="nav-item">
-                <a class="nav-link" data-toggle="tab" href="#notifications">{{ _("Notifications") }}
-                </a>
-            </li>
-        </ul>
-
-        <div class="border-bottom mb-4"></div>
-
-        <div class="tab-content">
-            <div class="tab-pane active" id="courses-enrolled" role="tabpanel" aria-labelledby="courses-enrolled">
-                {% include "lms/lms/web_template/courses_enrolled/courses_enrolled.html" %}
-            </div>
-
-            {% if show_creators_section %}
-            <div class="tab-pane fade" id="courses-created" role="tabpanel" aria-labelledby="courses-created">
-                {% include "lms/templates/courses_created.html" %}
-            </div>
-            {% endif %}
-
-            {% if show_review_section %}
-            <div class="tab-pane fade" id="courses-under-review" role="tabpanel" aria-labelledby="courses-under-review">
-                {% include "lms/templates/courses_under_review.html" %}
-            </div>
-            {% endif %}
-
-            <div class="tab-pane" id="notifications" role="tabpanel" aria-labelledby="notifications">
-                {% include "lms/templates/notifications.html" %}
-            </div>
-
-        </div>
-    </div>
-</div>
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-<script src="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-{% endblock %}
